Replace "PCIe PME" with "PCIe_PME" for /proc
In commit c7f486567c1d0acd2e4166c47069835b9f75e77b the driver was added,
with a space in the name.

This replaces the space with an underscore so that the directory in /proc
does not require quoting.

There are a couple-dozen more that could be changed as well:

```
git grep -l 'request_irq(.*".* .*".*)'
```

Co-authored-by: Brett Neumeier <brett@neumeier.us>

Signed-off-by: Eric Herman <eric@freesa.org>